+++ mactrack_convert.php
@@ -116,13 +116,13 @@ if (read_config_option('mt_collection_timing') != 'disabled') {
 	}
 
 	if ($partitioning == 'YES' || $partitioning == 'ACTIVE') {
-		mactrack_create_partitioned_table($engine, $days, true);
+		mactrack_create_partitioned_table($engine, $charset, $collate, $days, true);
 	} else {
 		print "FATAL: Partitioning Not Available, Exiting!\n";
 	}
 }
 
-function mactrack_create_partitioned_table($engine = 'InnoDB', $charset, $collate, $days = 30, $migrate = false) {
+function mactrack_create_partitioned_table($engine, $charset, $collate, $days = 30, $migrate = false) {
 	global $config;
 
 	// rename the original table