+++ mactrack_device_types.php
@@ -909,7 +909,7 @@ function mactrack_get_device_types(&$sql_where, $rows, $apply_limits = true) {
 	if (get_request_var('vendor') == 'All') {
 		// Show all items
 	} else {
-		$sql_where .= ($sql_where != '' ? ' AND ' : ' WHERE ') . "(mtdt.vendor='" . get_request_var('vendor') . "')";
+		$sql_where .= ($sql_where != '' ? ' AND ' : ' WHERE ') . 'mtdt.vendor = ' . db_qstr(get_request_var('vendor'));
 	}
 
 	if (get_request_var('type_id') == '-1') {